Add quarter, convert play time to s.ms

+++ basketball_reference/spiders/box_score/pbp.sql
@@ -13,7 +13,7 @@ CREATE TABLE pbp(
   make bool,
   orebounder TEXT,
   play TEXT,
-  play_time INT, /* m:ss -> seconds */
+  play_time REAL, /* m:ss:ms -> s.ms */
   play_time_raw TEXT,
   quarter INT,
   shooter TEXT,
